whereami/requirements.txt
flask
requests
emoji
flask-cors
grpcio
grpcio-reflection